Create cargo_manifest.sh
the printf "%d. %s\n" $((i + 1)) "${forward_bay[i]}" statement character by character:

printf: This is a command used to format and print data to the standard output.

"%d. %s\n": This is the format string that specifies how the output should be formatted:

%d: This is a format specifier that indicates an integer will be printed here. It will be replaced by the value of $((i + 1)).
.: This is a literal character that will be printed as-is, following the integer.
%s: This is a format specifier that indicates a string will be printed here. It will be replaced by the value of "${forward_bay[i]}".
\n: This is an escape sequence that represents a newline character. It moves the output to the next line after printing.
$((i + 1)): This is an arithmetic expansion that calculates the value of i + 1.

$(( ... )): This syntax is used for arithmetic operations in shell scripting.
i: This is the loop variable that represents the current index of the array.
+ 1: This adds 1 to the current index i, so that the numbering starts from 1 instead of 0.
"${forward_bay[i]}": This accesses the element of the forward_bay array at index i.

${...}: This syntax is used to reference the value of a variable or an array element.
forward_bay: This is the name of the array.
[i]: This specifies the index of the array element to access.
Putting it all together, the printf command formats the output to display the index (starting from 1) followed by a dot and the corresponding item from the forward_bay array, each on a new line.